# Conflicts:
#	Makefile
#	src/main/drivers/accgyro_spi_mpu6000.c
#	src/main/drivers/bus_i2c_stm32f30x.c
#	src/main/drivers/bus_spi.c
#	src/main/drivers/bus_spi.h
#	src/main/drivers/timer.c
#	src/main/flight/imu.h
#	src/main/io/serial_msp.c
#	src/main/mw.c

+++ src/main/drivers/accgyro_l3gd20.c
@@ -86,7 +86,7 @@ static void l3gd20SpiInit(SPI_TypeDef *SPIx)
     spiSetDivisor(L3GD20_SPI, MEDIUM_SPEED_SPI);
 }
 
-void l3gd20GyroInit(uint16_t lpf)
+void l3gd20GyroInit(uint8_t lpf)
 {
     UNUSED(lpf); // FIXME use it!
